chore(openaev): moving security platform to processor classes (#317)
