+++ openaev/openaev/processors/__init__.py
@@ -1,2 +1,7 @@
 from openaev.processors.json_api_processor import JsonApiProcessor
 from openaev.processors.json_flat_processor import JsonFlatProcessor
+
+__all_ = [
+    JsonApiProcessor,
+    JsonFlatProcessor,
+]
