+++ openaev/tests/test_default_expected_security_platforms.py
@@ -1,13 +1,11 @@
-from openaev.openaev_openaev import (
-    DEFAULT_EXPECTED_SECURITY_PLATFORMS,
-    apply_default_expected_security_platforms,
-)
+from openaev.constants import DEFAULT_EXPECTED_SECURITY_PLATFORMS
+from openaev.processors.common_processor import CommonProcessor
 
 
 def test_fills_defaults_for_declared_expectation_types():
     payload_information = {"payload_expectations": ["PREVENTION", "DETECTION"]}
 
-    apply_default_expected_security_platforms(payload_information)
+    CommonProcessor._apply_default_expected_security_platforms(payload_information)
 
     assert payload_information["payload_expected_security_platforms"] == {
         "DETECTION": ["EDR", "XDR", "SIEM"],
@@ -18,7 +16,7 @@ def test_fills_defaults_for_declared_expectation_types():
 def test_fills_defaults_only_for_declared_expectation_types():
     payload_information = {"payload_expectations": ["DETECTION"]}
 
-    apply_default_expected_security_platforms(payload_information)
+    CommonProcessor._apply_default_expected_security_platforms(payload_information)
 
     assert payload_information["payload_expected_security_platforms"] == {
         "DETECTION": ["EDR", "XDR", "SIEM"],
@@ -28,7 +26,7 @@ def test_fills_defaults_only_for_declared_expectation_types():
 def test_ignores_expectation_types_without_defaults():
     payload_information = {"payload_expectations": ["MANUAL"]}
 
-    apply_default_expected_security_platforms(payload_information)
+    CommonProcessor._apply_default_expected_security_platforms(payload_information)
 
     assert "payload_expected_security_platforms" not in payload_information
 
@@ -40,7 +38,7 @@ def test_explicit_value_takes_precedence():
         "payload_expected_security_platforms": explicit,
     }
 
-    apply_default_expected_security_platforms(payload_information)
+    CommonProcessor._apply_default_expected_security_platforms(payload_information)
 
     assert payload_information["payload_expected_security_platforms"] == {
         "DETECTION": ["SIEM"],
@@ -54,7 +52,7 @@ def test_explicit_empty_map_is_preserved():
         "payload_expected_security_platforms": {},
     }
 
-    apply_default_expected_security_platforms(payload_information)
+    CommonProcessor._apply_default_expected_security_platforms(payload_information)
 
     assert payload_information["payload_expected_security_platforms"] == {}
 
@@ -66,7 +64,7 @@ def test_explicit_null_is_treated_as_missing():
         "payload_expected_security_platforms": None,
     }
 
-    apply_default_expected_security_platforms(payload_information)
+    CommonProcessor._apply_default_expected_security_platforms(payload_information)
 
     assert payload_information["payload_expected_security_platforms"] == {
         "DETECTION": ["EDR", "XDR", "SIEM"],
@@ -76,7 +74,7 @@ def test_explicit_null_is_treated_as_missing():
 
 def test_payload_without_expectations_left_untouched():
     for payload_information in ({}, {"payload_expectations": []}):
-        apply_default_expected_security_platforms(payload_information)
+        CommonProcessor._apply_default_expected_security_platforms(payload_information)
 
         assert "payload_expected_security_platforms" not in payload_information
 
@@ -85,8 +83,8 @@ def test_defaults_are_copied_per_payload():
     first = {"payload_expectations": ["DETECTION"]}
     second = {"payload_expectations": ["DETECTION"]}
 
-    apply_default_expected_security_platforms(first)
-    apply_default_expected_security_platforms(second)
+    CommonProcessor._apply_default_expected_security_platforms(first)
+    CommonProcessor._apply_default_expected_security_platforms(second)
 
     first["payload_expected_security_platforms"]["DETECTION"].append("NDR")
